Arreglos terminados
- Corrección de arquitectura
- Bugs y dependencias inexistentes solucionadas

+++ app/src/main/java/com/example/transportistaapp/data/RepositoryImpl.kt
@@ -4,7 +4,6 @@ import android.util.Log
 import com.example.transportistaapp.data.database.dao.PaqueteDao
 import com.example.transportistaapp.data.database.dao.RutaDao
 import com.example.transportistaapp.data.database.entities.PaqueteEntity
-import com.example.transportistaapp.data.database.entities.RutaEntity
 import com.example.transportistaapp.data.database.entities.toDomain
 import com.example.transportistaapp.data.database.entities.toRoom
 import com.example.transportistaapp.data.network.FirestoreService
@@ -23,7 +22,7 @@ class RepositoryImpl @Inject constructor(
     private val rutaDao: RutaDao,
     private val paqueteDao: PaqueteDao
 
-) : Repository{
+) : Repository {
     override suspend fun loginTransportista(user: String, password: String): FirebaseUser? {
         try {
             val result = firebaseAuth.signInWithEmailAndPassword(user, password).await()
@@ -39,22 +38,28 @@ class RepositoryImpl @Inject constructor(
     }
 
     override suspend fun updateLocalPackages(transportista: String) {
+        if (rutaDao.rutasEnReparto().isNotEmpty()) {
+            return
+        }
         paqueteDao.deleteAll()
         rutaDao.deleteAll()
-        Log.d("Maybealog LoginTransportistaUseCase","Se  borran los datos")
         val rutasList = firestoreService.getRutasPorTransportista(transportista)
-        Log.d("Maybealog LoginTransportistaUseCase","Se obtiene")
-        Log.d("Maybealog LoginTransportistaUseCase",rutasList.toString())
+        Log.d("Maybealog LoginTransportistaUseCase", "Se obtiene")
+        Log.d("Maybealog LoginTransportistaUseCase", rutasList.toString())
         rutaDao.insertAll(rutasList.map {
-        Log.d("Maybealog LoginTransportistaUseCase",it.toRoom().toString())
+            Log.d("Maybealog LoginTransportistaUseCase", it.toRoom().toString())
             it.toRoom()
         })
-        Log.d("Maybealog LoginTransportistaUseCase","Se insertan las rutas")
+        Log.d("Maybealog LoginTransportistaUseCase", "Se insertan las rutas")
         rutasList.forEach { ruta ->
             val paquetesEntityList = ruta.paquetes.map { it.toRoom() }
             paqueteDao.insertAll(paquetesEntityList)
         }
-        Log.d("Maybealog LoginTransportistaUseCase","Se insertan los paquetes")
+        Log.d("Maybealog LoginTransportistaUseCase", "Se insertan los paquetes")
+    }
+
+    override suspend fun obtenerPaquetesEnReparto(): List<Paquete> {
+        return paqueteDao.obtenerEnReparto().map { it.toDomain() }
     }
 
     override suspend fun getRutasActivas(uid: String): List<Ruta> {
@@ -62,20 +67,20 @@ class RepositoryImpl @Inject constructor(
         return rutasFirebase
     }
 
-    override suspend fun marcarCajaEntregada(cajaId: String, fechaEntrega : Date) {
+    override suspend fun marcarCajaEntregada(cajaId: String, fechaEntrega: Date) {
         paqueteDao.entregar(cajaId, fechaEntrega)
     }
 
-    override suspend fun marcarCajaNoEntregada(cajaId: String) {
-        paqueteDao.entregaFallida(cajaId)
+    override suspend fun marcarCajaNoEntregada(cajaId: String, motivo: String) {
+        paqueteDao.entregaFallida(cajaId, motivo)
     }
 
-    override suspend fun obtenerPaquetesNoEntregados(): List<Paquete>? {
+    override suspend fun obtenerPaquetesNoEntregados(): List<Paquete> {
         val rutas = rutaDao.rutasEnReparto()
         val paquetesEntities = mutableListOf<PaqueteEntity>()
         rutas.forEach { paquetesEntities.addAll(paqueteDao.obtenerPorRuta(it.id)) }
         val paquetes = paquetesEntities.filter { it.estado >= 2 }.map { it.toDomain() }
-        return paquetes.ifEmpty { null }
+        return paquetes
     }
 
     override suspend fun terminarEntrega() {
@@ -84,7 +89,7 @@ class RepositoryImpl @Inject constructor(
         rutas.forEach {
             paquetes.addAll(paqueteDao.obtenerPorRuta(it.id))
         }
-        firestoreService.terminarRutas(rutas.map {it.toDomain()})
+        firestoreService.terminarRutas(rutas.map { it.toDomain() })
         firestoreService.terminarPaquetes(paquetes.map { it.toDomain() })
     }
